requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==4.2
django-environ==0.11.2
django-seed==0.3.1
djangorestframework==3.15.2
Faker==28.4.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.1
